Make fzy appear under the prompt, instead of over it
This is achieved by inserting a call to "zle -M", which inserts a line
below the prompt, but remembers the location of the prompt before the
widget was triggered and returns to that position after the widget is
exited. Typically this is used to write e.g. completion hints below the
prompt, but here we just abuse it to force Zsh to move the cursor one
line below the prompt, effectively making fzy show under the prompt.
The string "<fzy>" passed is dummy, and will be immediately overwritten
by fzy.

Fixes #12